+++ .idea/gradle.xml
@@ -9,6 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/annotation" />
+            <option value="$PROJECT_DIR$/processor" />
           </set>
         </option>
       </GradleProjectSettings>
