+++ annotation/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    kotlin("jvm")
+}
+
+group = "org.example"
+version = "1.0-SNAPSHOT"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(kotlin("stdlib"))
+}
