+++ settings.gradle.kts
@@ -1,2 +1,3 @@
 rootProject.name = "kspexample"
-
+include("annotation")
+include("processor")
