+++ src/main/kotlin/com/tsongkha/kspexample/Main.kt
@@ -1,0 +1,5 @@
+package com.tsongkha.kspexample
+
+fun main() {
+    println("hello")
+}
